# Fix `custom-property-no-missing-var-function` performance