--- a/.changeset/heavy-pandas-cheer.md
+++ b/.changeset/heavy-pandas-cheer.md
@@ -1,0 +1,5 @@
+---
+"stylelint": patch
+---
+
+Fixed: `custom-property-no-missing-var-function` performance