--- a/lib/rules/custom-property-no-missing-var-function/index.js
+++ b/lib/rules/custom-property-no-missing-var-function/index.js
@@ -3,7 +3,6 @@
 const valueParser = require('postcss-value-parser');
 
 const declarationValueIndex = require('../../utils/declarationValueIndex');
-const isCustomProperty = require('../../utils/isCustomProperty');
 const report = require('../../utils/report');
 const ruleMessages = require('../../utils/ruleMessages');
 const validateOptions = require('../../utils/validateOptions');
@@ -17,6 +16,9 @@ const messages = ruleMessages(ruleName, {
 const meta = {
 	url: 'https://stylelint.io/user-guide/rules/custom-property-no-missing-var-function',
 };
+
+const DOUBLE_DASH = /--/;
+const DASHED_IDENT_START = /^--/;
 
 /** @type {import('stylelint').Rule} */
 const rule = (primary) => {
@@ -32,12 +34,15 @@ const rule = (primary) => {
 			knownCustomProperties.add(atRule.params);
 		});
 
-		root.walkDecls(({ prop }) => {
-			if (isCustomProperty(prop)) knownCustomProperties.add(prop);
+		root.walkDecls(DASHED_IDENT_START, ({ prop }) => {
+			knownCustomProperties.add(prop);
 		});
 
 		root.walkDecls((decl) => {
 			const { value } = decl;
+
+			if (!DOUBLE_DASH.test(value)) return;
+
 			const parsedValue = valueParser(value);
 
 			parsedValue.walk((node) => {