--- a/lib/rules/custom-property-no-missing-var-function/index.js
+++ b/lib/rules/custom-property-no-missing-var-function/index.js
@@ -17,8 +17,8 @@ const meta = {
 	url: 'https://stylelint.io/user-guide/rules/custom-property-no-missing-var-function',
 };
 
-const DOUBLE_DASH = /--/;
 const DASHED_IDENT_START = /^--/;
+const DOUBLE_DASH = /--/;
 
 /** @type {import('stylelint').Rule} */
 const rule = (primary) => {