--- a/lib/rules/custom-property-no-missing-var-function/index.js
+++ b/lib/rules/custom-property-no-missing-var-function/index.js
@@ -17,9 +17,6 @@ const meta = {
 	url: 'https://stylelint.io/user-guide/rules/custom-property-no-missing-var-function',
 };
 
-const DASHED_IDENT_START = /^--/;
-const DOUBLE_DASH = /--/;
-
 /** @type {import('stylelint').Rule} */
 const rule = (primary) => {
 	return (root, result) => {
@@ -34,14 +31,14 @@ const rule = (primary) => {
 			knownCustomProperties.add(atRule.params);
 		});
 
-		root.walkDecls(DASHED_IDENT_START, ({ prop }) => {
+		root.walkDecls(/^--/, ({ prop }) => {
 			knownCustomProperties.add(prop);
 		});
 
 		root.walkDecls((decl) => {
 			const { value } = decl;
 
-			if (!DOUBLE_DASH.test(value)) return;
+			if (!value.includes('--')) return;
 
 			const parsedValue = valueParser(value);