<!-- Each pull request must be associated with an open issue unless it's a documentation fix. If a corresponding issue does not exist, please create one so we can discuss the change first. -->

<!-- Please answer the following. We close pull requests that don't. -->

> Which issue, if any, is this issue related to?

See : https://github.com/stylelint/stylelint/issues/6869

> Is there anything in the PR that needs further explanation?

No, it's self-explanatory.

----

before :

```
Warnings: 0
Mean: 106.64779905882351 ms
Deviation: 14.50380702371365 ms
```

after :

```
Warnings: 0
Mean: 98.50316917647058 ms
Deviation: 11.405372458359288 ms
```